Add function browser tests.
