+++ js-api-spec/value/map.test.ts
@@ -5,8 +5,6 @@
 import {Value, SassList, SassMap, SassNumber, SassString} from 'sass';
 import {List, OrderedMap} from 'immutable';
 
-import '../utils';
-
 describe('SassMap', () => {
   let map: SassMap;
   beforeEach(() => {
